choir: unify duplicated sproof generation code on Subscriber
