+++ cumulus/pallets/subscriber/src/benchmarking.rs
@@ -18,65 +18,11 @@
 #![cfg(feature = "runtime-benchmarks")]
 
 use super::*;
-use crate::Pallet as Subscriber;
-use codec::Encode;
-use cumulus_pallet_parachain_system::RelayChainStateProof;
+use crate::{mock::build_sproof_with_child_data, Pallet as Subscriber};
 use cumulus_primitives_core::ParaId;
 use frame_benchmarking::v2::*;
 use frame_support::traits::Get;
 use frame_system::RawOrigin;
-use sp_runtime::{traits::HashingFor, StateVersion};
-use sp_state_machine::{Backend, TrieBackendBuilder};
-use sp_trie::PrefixedMemoryDB;
-
-/// Build a relay chain state proof with child trie data for multiple publishers.
-fn build_test_proof<T: Config>(
-	publishers: &[(ParaId, Vec<(Vec<u8>, Vec<u8>)>)],
-) -> RelayChainStateProof {
-	let (db, root) = PrefixedMemoryDB::<HashingFor<polkadot_primitives::Block>>::default_with_root();
-	let state_version = StateVersion::default();
-	let mut backend = TrieBackendBuilder::new(db, root).build();
-
-	let mut all_proofs = vec![];
-	let mut main_trie_updates = vec![];
-
-	// Process each publisher
-	for (publisher_para_id, child_data) in publishers {
-		let child_info = sp_core::storage::ChildInfo::new_default(&(b"pubsub", *publisher_para_id).encode());
-
-		// Insert child trie data
-		let child_kv: Vec<_> = child_data.iter().map(|(k, v)| (k.clone(), Some(v.clone()))).collect();
-		backend.insert(vec![(Some(child_info.clone()), child_kv)], state_version);
-
-		// Get child trie root and prepare to insert it in main trie
-		let child_root = backend.child_storage_root(&child_info, core::iter::empty(), state_version).0;
-		let prefixed_key = child_info.prefixed_storage_key();
-		main_trie_updates.push((prefixed_key.to_vec(), Some(child_root.encode())));
-
-		// Prove child trie keys
-		let child_keys: Vec<_> = child_data.iter().map(|(k, _)| k.clone()).collect();
-		if !child_keys.is_empty() {
-			let child_proof = sp_state_machine::prove_child_read_on_trie_backend(&backend, &child_info, child_keys)
-				.expect("prove child read");
-			all_proofs.push(child_proof);
-		}
-	}
-
-	// Insert all child roots in main trie
-	backend.insert(vec![(None, main_trie_updates.clone())], state_version);
-	let root = *backend.root();
-
-	// Prove all child roots in main trie
-	let main_keys: Vec<_> = main_trie_updates.iter().map(|(k, _)| k.clone()).collect();
-	let main_proof = sp_state_machine::prove_read_on_trie_backend(&backend, main_keys)
-		.expect("prove read");
-	all_proofs.push(main_proof);
-
-	// Merge all proofs
-	let proof = sp_trie::StorageProof::merge(all_proofs);
-
-	RelayChainStateProof::new(ParaId::from(100), root, proof).expect("valid proof")
-}
 
 /// Create test subscriptions for benchmarking.
 fn create_subscriptions(n: u32, keys_per_publisher: u32) -> Vec<(ParaId, Vec<Vec<u8>>)> {
@@ -123,7 +69,7 @@ mod benchmarks {
 		let publishers: Vec<_> = (0..n)
 			.map(|i| (ParaId::from(1000 + i), vec![(vec![i as u8], vec![25u8])]))
 			.collect();
-		let proof = build_test_proof::<T>(&publishers);
+		let proof = build_sproof_with_child_data(&publishers);
 
 		#[block]
 		{
@@ -153,7 +99,7 @@ mod benchmarks {
 				(para_id, child_data)
 			})
 			.collect();
-		let proof = build_test_proof::<T>(&publishers);
+		let proof = build_sproof_with_child_data(&publishers);
 		let current_roots = Subscriber::<T>::collect_publisher_roots(&proof, &subscriptions);
 
 		#[block]
